requirements.txt
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-contrib-python>=4.1.2.30
pillow
PyYAML>=5.3
scipy>=1.4.1
tqdm>=4.41.0
flask-cors
gunicorn
torch
torchvision
pandas
ipython
psutil
seaborn